Brain_Tumor_AI_Chatbot_Cloud/requirements.txt
streamlit==1.39.0
tensorflow-cpu==2.17.0
opencv-python==4.10.0.84
numpy==1.26.4
pandas==2.2.2
matplotlib==3.8.4
Pillow==10.4.0
scikit-learn==1.5.2
requests==2.32.3
